Remove code climate bunder audit because it fails
CodeClimiate retunrs an error during the bundler audit step, because it
does not have a Gemfile.lock file.

Since I do not want to checking in the Gemfile.lock because this is a
gem, I am going the remove this plugin. If this behaviour changes in
the future, then I will try it again.